+++ metricbeat/docs/modules/ceph.asciidoc
@@ -19,7 +19,7 @@ in <<configuration-metricbeat>>. Here is an example configuration:
 ----
 metricbeat.modules:
 #- module: ceph
-#  metricsets: ["health", "monitor"]
+#  metricsets: ["cluster_health", "monitor_health"]
 #  enabled: true
 #  period: 10s
 #  hosts: ["localhost:5000"]
@@ -30,11 +30,11 @@ metricbeat.modules:
 
 The following metricsets are available:
 
-* <<metricbeat-metricset-ceph-cluster,cluster>>
+* <<metricbeat-metricset-ceph-cluster_health,cluster_health>>
 
-* <<metricbeat-metricset-ceph-monitor,monitor>>
+* <<metricbeat-metricset-ceph-monitor_health,monitor_health>>
 
-include::ceph/cluster.asciidoc[]
+include::ceph/cluster_health.asciidoc[]
 
-include::ceph/monitor.asciidoc[]
+include::ceph/monitor_health.asciidoc[]
