+++ metricbeat/docs/modules/ceph/cluster_health.asciidoc
@@ -2,8 +2,8 @@
 This file is generated! See scripts/docs_collector.py
 ////
 
-[[metricbeat-metricset-ceph-cluster]]
-include::../../../module/ceph/cluster/_meta/docs.asciidoc[]
+[[metricbeat-metricset-ceph-cluster_health]]
+include::../../../module/ceph/cluster_health/_meta/docs.asciidoc[]
 
 
 ==== Fields
@@ -15,5 +15,5 @@ Here is an example document generated by this metricset:
 
 [source,json]
 ----
-include::../../../module/ceph/cluster/_meta/data.json[]
+include::../../../module/ceph/cluster_health/_meta/data.json[]
 ----
