+++ metricbeat/docs/modules/ceph/monitor_health.asciidoc
@@ -2,8 +2,8 @@
 This file is generated! See scripts/docs_collector.py
 ////
 
-[[metricbeat-metricset-ceph-monitor]]
-include::../../../module/ceph/monitor/_meta/docs.asciidoc[]
+[[metricbeat-metricset-ceph-monitor_health]]
+include::../../../module/ceph/monitor_health/_meta/docs.asciidoc[]
 
 
 ==== Fields
@@ -15,5 +15,5 @@ Here is an example document generated by this metricset:
 
 [source,json]
 ----
-include::../../../module/ceph/monitor/_meta/data.json[]
+include::../../../module/ceph/monitor_health/_meta/data.json[]
 ----
