+++ metricbeat/include/list.go
@@ -11,8 +11,8 @@ import (
 	_ "github.com/elastic/beats/metricbeat/module/apache"
 	_ "github.com/elastic/beats/metricbeat/module/apache/status"
 	_ "github.com/elastic/beats/metricbeat/module/ceph"
-	_ "github.com/elastic/beats/metricbeat/module/ceph/cluster"
-	_ "github.com/elastic/beats/metricbeat/module/ceph/monitor"
+	_ "github.com/elastic/beats/metricbeat/module/ceph/cluster_health"
+	_ "github.com/elastic/beats/metricbeat/module/ceph/monitor_health"
 	_ "github.com/elastic/beats/metricbeat/module/couchbase"
 	_ "github.com/elastic/beats/metricbeat/module/couchbase/bucket"
 	_ "github.com/elastic/beats/metricbeat/module/couchbase/cluster"
