+++ metricbeat/metricbeat.full.yml
@@ -96,7 +96,7 @@ metricbeat.modules:
 
 #-------------------------------- ceph Module --------------------------------
 #- module: ceph
-#  metricsets: ["health", "monitor"]
+#  metricsets: ["cluster_health", "monitor_health"]
 #  enabled: true
 #  period: 10s
 #  hosts: ["localhost:5000"]
