+++ metricbeat/metricbeat.template-es2x.json
@@ -186,7 +186,7 @@
         },
         "ceph": {
           "properties": {
-            "cluster": {
+            "cluster_health": {
               "properties": {
                 "overall_status": {
                   "ignore_above": 1024,
@@ -214,7 +214,7 @@
                 }
               }
             },
-            "monitor": {
+            "monitor_health": {
               "properties": {
                 "available": {
                   "properties": {
