--- metricbeat/module/ceph/cluster/_meta/docs.asciidoc
@@ -1,3 +1,0 @@
-=== ceph health MetricSet
-
-This is the health metricset of the module ceph.
