+++ metricbeat/module/ceph/cluster_health/_meta/docs.asciidoc
@@ -1,0 +1,3 @@
+=== ceph cluster_health MetricSet
+
+This is the cluster_health metricset of the module ceph.
