+++ metricbeat/module/ceph/monitor_health/_meta/docs.asciidoc
@@ -1,0 +1,3 @@
+=== ceph monitor_health MetricSet
+
+This is the monitor_health metricset of the module ceph.
