+++ metricbeat/module/ceph/monitor_health/monitor_health_integration_test.go
@@ -1,4 +1,4 @@
-package cluster
+package monitor_health
 
 import (
 	"fmt"
@@ -9,8 +9,8 @@ import (
 )
 
 func TestData(t *testing.T) {
-	f := mbtest.NewEventFetcher(t, getConfig())
-	err := mbtest.WriteEvent(f, t)
+	f := mbtest.NewEventsFetcher(t, getConfig())
+	err := mbtest.WriteEvents(f, t)
 	if err != nil {
 		t.Fatal("write", err)
 	}
@@ -19,7 +19,7 @@ func TestData(t *testing.T) {
 func getConfig() map[string]interface{} {
 	return map[string]interface{}{
 		"module":     "ceph",
-		"metricsets": []string{"cluster"},
+		"metricsets": []string{"monitor_health"},
 		"hosts":      getTestCephHost(),
 	}
 }
